--- test/solutions/chapter_5/subchapter_5/lesson_3/s_5_5_3_1_test.dart
@@ -1,9 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  testWidgets('S5531', (tester) async {
-    // Testen dieser Aufgabe verlangt Dependency Injection und Mocking.
-    // das würde leider auch den Setup-Code für die TN komplizierter machen und zu Verwirrung führen.
-    // deshalb wird diese Aufgabe nicht getestet
-  });
-}
